requirements.txt
PySide6
Pillow
exifread
